Python Changes:  - changed value type in python from float to double  - added AlphaExpasionFusion to Python

+++ src/interfaces/python/examples/inference_bp.py
@@ -1,7 +1,7 @@
 import opengm
 import numpy
 
-unaries=numpy.random.rand(10, 10,2).astype(numpy.float32)
+unaries=numpy.random.rand(10, 10,2)
 potts=opengm.PottsFunction([2,2],0.0,0.4)
 gm=opengm.grid2d2Order(unaries=unaries,regularizer=potts)
